requirements.txt
aiohttp
botbuilder-core
botbuilder-schema
pandas
openpyxl
requests